requirements.txt
# Core agentic framework
langchain>=0.3.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# LLM providers — install the one(s) you need
langchain-groq>=0.2.0          # Groq (free, Llama 3.3 70B) — recommended
langchain-ollama>=0.2.0        # Ollama local models
langchain-anthropic>=0.3.0     # Claude Sonnet (paid)

# Market data & analysis
yfinance>=0.2.40
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.0
ta>=0.11.0

# UI & visualisation
streamlit>=1.40.0
plotly>=5.18.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
beautifulsoup4>=4.12.0
